Initial refactor of views.py
